Extract SigningService out of ComplianceService
