+++ api/src/org/labkey/api/compliance/ComplianceService.java
@@ -17,9 +17,6 @@ package org.labkey.api.compliance;
 
 import org.jetbrains.annotations.NotNull;
 import org.json.JSONObject;
-import org.labkey.api.attachments.AttachmentParent;
-import org.labkey.api.attachments.AttachmentType;
-import org.labkey.api.attachments.ByteArrayAttachmentFile;
 import org.labkey.api.data.Activity;
 import org.labkey.api.data.Container;
 import org.labkey.api.data.PHI;
@@ -28,16 +25,9 @@ import org.labkey.api.query.column.ColumnInfoFilter;
 import org.labkey.api.query.column.ColumnInfoTransformer;
 import org.labkey.api.security.User;
 import org.labkey.api.services.ServiceRegistry;
-import org.labkey.api.util.Pair;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.UnauthorizedException;
 import org.labkey.api.view.ViewContext;
-
-import java.io.IOException;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.Map;
-import java.util.function.Function;
 
 /**
  * Created by davebradlee on 7/27/17.
@@ -65,9 +55,14 @@ public interface ComplianceService
 
     boolean isComplianceSupported();
     String getModuleName();
+
+    // Next three methods are signing-related, BUT the implementation is closely tied to the compliance controller
+    // and the functionality is part of our full compliance capabilities. (As opposed to ELN which just needs the
+    // signing service.) So they remain on ComplianceService.
     ActionURL urlFor(Container container, QueryAction action, ActionURL queryBasedUrl);
     boolean hasElecSignPermission(@NotNull Container container, @NotNull User user);
     boolean hasViewSignedSnapshotsPermission(@NotNull Container container, @NotNull User user);
+
     default @NotNull PHI getMaxAllowedPhi(@NotNull Container container, @NotNull User user)
     {
         return PHI.Restricted;
@@ -88,23 +83,6 @@ public interface ComplianceService
      * @throws org.labkey.api.view.UnauthorizedException if the user doesn't have sufficient permissions.
      */
     void setFolderSettings(@NotNull Container container, @NotNull User user, @NotNull ComplianceFolderSettings settings);
-
-    /**
-     * Register a supplier for an attachment parent for attachment types that should use an attachment parent that differs
-     * from the default one provided (for example: saving the attachment to the file system instead of the DB).
-     * The attachment type will need to be specified on the SignedSnapshot.
-     */
-    void registerAttachmentParentSupplier(AttachmentType type, Function<SignedSnapshot, AttachmentParent> supplier);
-
-    /**
-     * CRD operations for signed snapshots
-     */
-    Integer insertSignedSnapshot(Container container, User user, SignedSnapshot snapshot, ByteArrayAttachmentFile file) throws IOException;
-    SignedSnapshot getSignedSnapshot(Container container, int snapshotId);
-    Collection<SignedSnapshot> getSignedSnapshots(Container container, String ownerEntityId);
-    Map<String, Object> deleteSignedSnapshot(Container container, User user, int snapshotId);
-
-    Pair<AttachmentParent, String> downloadSnapshot(Container container, User user, @NotNull SignedSnapshot snapshot);
 
     JSONObject getPageContextJson();
 
@@ -160,41 +138,6 @@ public interface ComplianceService
         @Override
         public void setFolderSettings(@NotNull Container container, @NotNull User user, @NotNull ComplianceFolderSettings settings)
         {
-        }
-
-        @Override
-        public Integer insertSignedSnapshot(Container container, User user, SignedSnapshot snapshot, ByteArrayAttachmentFile file) throws IOException
-        {
-            return null;
-        }
-
-        @Override
-        public void registerAttachmentParentSupplier(AttachmentType type, Function<SignedSnapshot, AttachmentParent> supplier)
-        {
-        }
-
-        @Override
-        public SignedSnapshot getSignedSnapshot(Container container, int snapshotId)
-        {
-            return null;
-        }
-
-        @Override
-        public Collection<SignedSnapshot> getSignedSnapshots(Container container, String ownerEntityId)
-        {
-            return Collections.emptyList();
-        }
-
-        @Override
-        public Map<String, Object> deleteSignedSnapshot(Container container, User user, int snapshotId)
-        {
-            return null;
-        }
-
-        @Override
-        public Pair<AttachmentParent, String> downloadSnapshot(Container container, User user, SignedSnapshot snapshot)
-        {
-            return null;
         }
 
         @Override
